fix: reset sync when another retry happens before synced
